Reorganise scripts for clarity.
I found myself typing 'update' which autocompleted to
'update-plugins.sh', when in actual fact the 'install.sh' script is the
one that handles routing updating.

So I broke out the 'install' and 'update' steps into separate scripts,
and renamed the plugin updater to 'fetch-latest-plugins.sh' to hopefully
make it clearer that this is a maintenence script rather than a routine
update thing.